+++ common/src/main/java/io/github/flemmli97/mobbattle/MobBattle.java
@@ -1,6 +1,10 @@
 package io.github.flemmli97.mobbattle;
 
 import io.github.flemmli97.tenshilib.TenshiLib;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.EquipmentSlot;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -16,6 +20,8 @@ public class MobBattle {
 
     public static final EquipmentSlot[] slot = {EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND, EquipmentSlot.HEAD,
             EquipmentSlot.CHEST, EquipmentSlot.LEGS, EquipmentSlot.FEET};
+
+    public static TagKey<EntityType<?>> IGNORED = TagKey.create(Registry.ENTITY_TYPE_REGISTRY, new ResourceLocation(MobBattle.MODID, "ignored_mobs"));
 
     @SuppressWarnings("unchecked")
     public static <T> T getPlatformInstance(Class<T> abstractClss, String fabricImpl, String forgeImpl) {