+++ forge/src/main/java/io/github/flemmli97/mobbattle/forge/handler/EventHandler.java
@@ -11,6 +11,8 @@ import net.minecraft.world.entity.monster.Vex;
 import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.event.entity.living.LivingAttackEvent;
 import net.minecraftforge.event.entity.living.LivingEvent;
+import net.minecraftforge.event.entity.living.LivingHurtEvent;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 public class EventHandler {
@@ -36,6 +38,15 @@ public class EventHandler {
             if (Utils.isOnSameTeam(event.getEntity(), attacker) && !event.getEntity().getTeam().isAllowFriendlyFire())
                 event.setCanceled(true);
         }
+    }
+
+    /**
+     * Vanilla sets it in Mob#doHurtTarget but all mobs that override that method dont so...
+     */
+    @SubscribeEvent(priority = EventPriority.LOWEST)
+    public void setHurtMob(LivingHurtEvent event) {
+        if (event.getSource().getEntity() instanceof Mob mob)
+            mob.setLastHurtMob(event.getEntity());
     }
 
     @SubscribeEvent